requirements.txt
# Core dependencies for QuantumQA Autonomous QA Agent
# Python 3.11+ required

# Web UI Framework
streamlit>=1.28.0

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.2

# Document Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=2.0.0
pymupdf>=1.23.0

# LLM Integration (switched from OpenAI to Google Gemini)
google-generativeai>=0.3.0

# Selenium for Test Script Generation
selenium>=4.15.0

# Data & Utilities
pandas>=2.0.0
pydantic>=2.0.0
markdown>=3.5.0

# LangChain for RAG
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Environment variable loading
python-dotenv>=1.0.0

# ML Framework (CPU-optimized for Streamlit Cloud)
torch>=2.0.0; sys_platform != 'win32'
torch>=2.0.0; sys_platform == 'win32'